requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-filter==2.4.0
djangorestframework==3.12.2
Markdown==3.3.3
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.4.1
